Added the two required questions, added the repr library
Added a broad question: Which "kinds" of players are most likely to contribute a large amount of data so that we can target those players in our recruiting efforts.

Added a specific question:
How well do players’ total played hours and total number of sessions predict the amount of gameplay data they contribute?

and added the repr library